Move calculate_damage/modifier into calculate
